renamed messageinput into typingbar, moved into its own folder in components for organization, made new typingbar.css to prep for styling
