+++ frontend/src/components/typingbar/Typingbar.js
